packages/tiny-utils-fp/src/stream/ChildProcessStreamPipe.test.ts
/* eslint-disable fp/no-mutating-methods,fp/no-throw */
import * as childProcess from 'node:child_process';
import { PassThrough, Readable } from 'node:stream';

import { BufferWriteableStream } from './BufferWriteableStream';
import * as unit from './ChildProcessStreamPipe';
import { waitForWriteStreamPromise } from './index';

describe('ChildProcessStreamPipe', () => {
  const TEST_S = 'hoi noi broin coi' as const;

  it('should formulate args as expected', () => {
    const actual = new unit.ChildProcessStreamPipe('sed', []);
    actual.arg('-foo bar').arg('-baz');

    expect(actual.args).toStrictEqual(['-foo', 'bar', '-baz']);
  });

  it('should work as expected in the success case', async () => {
    const actual = new unit.ChildProcessStreamPipe('sed', ['-e', 's/\\(.*\\)/\\U\\1/']).input('-').output('-');
    const inStream = Readable.from(TEST_S);
    const outStream = new BufferWriteableStream();

    inStream.pipe(actual).pipe(outStream);
    await waitForWriteStreamPromise(outStream);
    actual._flush(jest.fn);

    expect(outStream.string).toEqual('HOI NOI BROIN COI');
  });

  // Test fails due to Unhandled error
  // FIXME: Needs further work
  xit('should work as expected with a stdin error', async () => {
    const actual = new unit.ChildProcessStreamPipe('sed', ['-e', 's/\\(.*\\)/\\U\\1/']).input('-').output('-');
    const inStream = Readable.from(TEST_S);
    const outStream = new BufferWriteableStream();

    inStream.pipe(actual).pipe(outStream);
    actual._childProcess?.stdin.emit('error', new Error('BOOM!'));
    await waitForWriteStreamPromise(outStream);

    expect(outStream.string).toEqual('');
  });

  it('should work as expected with a stdin EPIPE error', async () => {
    const actual = new unit.ChildProcessStreamPipe('sed', ['-e', 's/\\(.*\\)/\\U\\1/']).input('-').output('-');
    const inStream = Readable.from(TEST_S);
    const outStream = new BufferWriteableStream();

    inStream.pipe(actual).pipe(outStream);
    actual._childProcess?.stdin.emit('error', { code: 'EPIPE' });
    await waitForWriteStreamPromise(outStream);

    expect(outStream.string).toEqual('');
  });

  // Test fails with: TypeError: Cannot redefine property: spawn
  // FIXME: Needs more work
  xit('should work as expected in error case', () => {
    jest.spyOn(childProcess, 'spawn').mockImplementationOnce(() => {
      throw new Error('BOOM!');
    });
    const inStream = Readable.from(TEST_S);
    const outStream = new PassThrough();
    const actual = new unit.ChildProcessStreamPipe('cat', ['/tmp/LICENSE']);
    inStream.pipe(actual).pipe(outStream);

    expect(() => {
      return waitForWriteStreamPromise(outStream);
    }).toBeDefined();
  });

  // Causes Unhandled error, but still passes.
  // Plus is not really a proper test for errors due to tricky stream stuff
  // FIXME: Needs further work
  xit('should work as expected in error case', () => {
    const inStream = Readable.from(TEST_S);
    const outStream = new PassThrough();
    const actual = new unit.ChildProcessStreamPipe('cat', ['/tmp/doesnotexist']);
    inStream.pipe(actual).pipe(outStream);

    expect(() => {
      return waitForWriteStreamPromise(outStream);
    }).toBeDefined();
  });
});
